move branch lookup logic to helper
